+++ project/src/constants/Sizes.ts
@@ -1,0 +1,8 @@
+export default {
+    XXL: '2rem',
+    XL: '1.6rem',
+    L: '1.4rem',
+    M: '1rem',
+    S: '0.8rem',
+    XS: '0.5rem',
+};